Allow host app & extensions to add Tailwind CSS code
In a similar way we were doing for the tailwind configuration file, we're
now allowing the host app and extensions to add their own Tailwind CSS.

This is done by changing the `application.tailwind.css` file to an ERB
template. Both the host application and extensions can add their own
stylesheets through the `tailwind_stylesheets` configuration option. During
compilation time, the content for those `tailwind_stylesheets` is appended to
the dynamically generated CSS file.

As adding new styles should be thoroughly thought as per tailwind recomendations,
we're not creating a default stylesheet for the host app and, instead, it needs
to be manually created.

The way we're doing this can be seen as a bit clumsy, but the alternative
requires using `postcss-import` and adding all the npm jazz to the mix. There have
been some discussion on tailwindcss-rails about this [2], but there's no built-in way as of
today.

[1] - https://tailwindcss.com/docs/using-with-preprocessors#build-time-imports
[2] - https://github.com/rails/tailwindcss-rails/issues/133