Reset Taxon Paperclip attributes on attachment destroy
When using Paperclip, calling destroy on the attachment removes it from
from the storage location, such as S3, and leaves dirty changes on the
Paperclip attributes, such as icon_file_name.

If a store used `.present?` on the icon, it would return continue to
return `true`, as it looks at the file_name attribute [1], yet the
icon no longer exists on storage (`.exists? => false`).

A view such as the backend Taxon edit view, could therefore attempt to
show the icon, but it would fail and show a broken icon image.

Note:
- an assertion was not added for `.present?` as ActiveStorage's API
  works a little differently;
- stores using Paperclip may have Taxons with Paperclip attributes
  present, yet the image on storage no longer exists.

[1] https://github.com/kreeti/kt-paperclip/blob/91923a6ee1bd692696b51ea839f2f2578cb485b4/lib/paperclip/attachment.rb#L357-L361