Add support for themes to solidus_backend
The default theme is named "classic" and will refer back to
`spree/backend/all` so that any existing customization will keep
working.

Additional core themes should live under `spree/backend/themes/` and
be added to the backend preference. Custom themes can be added as well
rewriting or augmenting an existing theme.

This is all preparatory work for the new solidus_admin theme and to
allow existing stores do a smooth and controlled transition to it.