Ensure to return false on Paperclip destroy_attachment failure
If the attachment does exist, destroying the attachment will be
attempted. However, destroying is not guaranteed by Paperclip. In case
of failure, it will return `false` [1][2]. It should therefore return
`false` in this case, so callees know it failed. Before, due to
the `save` call afterwards, it was always returning `true`.

[1] https://github.com/kreeti/kt-paperclip/blob/v7.1.1/lib/paperclip/attachment.rb#L263
[2] https://github.com/kreeti/kt-paperclip/blob/v7.1.1/lib/paperclip/attachment.rb#L234-L235

Co-authored-by: Andrea Longhi <andrea@spaghetticode.it>