Fix development version
Currently, the update draft release task is failing on v4 [1] because it
looks like [2] a patch branch (`v4`) is tracking a next major release
(`v4.0.0`).

The underlying issue is that we don't bump to the development version
when we create a new patch branch [3]. To fix that for next releases, we
need to create another PR on the newly created patch branch.

The issue doesn't affect when bumping to the next release, as the
candidate version is taken from the list of git tags [4], and not from the
version file.

[1] - https://github.com/solidusio/solidus/actions/runs/5086912864
[2] - https://github.com/solidusio/solidus/blob/474877aaad9139a0a9b8564a9e7279e5446bc12c/dev_tools/lib/solidus/pipeline_context.rb#L34
[3] - https://github.com/solidusio/solidus/blob/474877aaad9139a0a9b8564a9e7279e5446bc12c/.github/workflows/prepare_post_release.yml#L62
[4] - https://github.com/solidusio/solidus/blob/474877aaad9139a0a9b8564a9e7279e5446bc12c/.github/actions/extract_pipeline_context/action.yml#L40