Allow changing the order recalculator
This adds a preference to change the order updater. The order updater, a
very central piece of Solidus' infrastructure, carries a bit of
technical debt. Unfortunately, it's extremely hard to change because
every line of this class might have been overwritten by implementors.

This configuration allows switching the order updater class, with the
intention of being able to introduce breaking changes from a gem without
patching the original order recalculator.

Additionally, the `order.recalculate` and `order.recalculator` are
introduced as aliases to the now deprecated `order.update` and
`order.updater` in order to clarify the intention of these methods.