Bump to the dev version when a new patch branch is created
When there's an existing patch branch (e.g., `v3.4`, `v4.0`...) and a
new release is made, we're already bumping to the `.dev` version for the
next patch release [1].

However, when a new patch branch is created [2] we were missing this.
For instance, after releasing `v4.0.0` the newly created `v4.1` branch
would still point at `4.0.0` instead of `4.0.1.dev` in `Spree::VERSION`.

That wasn't a problem for minor releases, but it caused issues when a
patch branch looked like tracking a next major release (like the given
`4.0.0` example) [3].

We're fixing this by creating another Pull Request for the new created
patch branch, bumping to the next patch `.dev` version. To allow this,
we're modifying the pull request action to allow specifying a base
branch different than the one running the workflow, and allowing a
workflow to create more than one pull request by adding a timestamp to
the branches it generates.

[1] - https://github.com/solidusio/solidus/blob/3d9acfcc83cc67d91db6e1573247a14b28329d75/.github/workflows/prepare_post_release.yml#L29
[2] - https://github.com/solidusio/solidus/blob/3d9acfcc83cc67d91db6e1573247a14b28329d75/.github/workflows/prepare_post_release.yml#L62
[3] - https://github.com/solidusio/solidus/pull/5108