Enhance Role model for better role management
This commit expands the functionality of the `Spree::Role` model to provide more
granular and efficient control over user roles and permissions.

  * Spree::RolePermission
  * Spree::Role enhancements
  * Spree::PermissionSet

These enhancements are a part of our broader initiative to improve the
flexibility and extensibility of role and permission management in Solidus,
adapting functionality from the `solidus_user_roles` gem.