Use a placeholder CSS file for solidus_admin testing
Solidus Admin generates the compiled tailwind CSS file in the host
application. For now, we follow a similar pattern to the one used for
other Solidus components and add a placeholder CSS file to the dummy
application. It'd be good to find a way to avoid the dummy app knowing
about its consumers.