Delete new minor release issue template (moved to .github)
Unfortunately, if we create a new issue template on this repository, GitHub immediately stops
inheriting the other issue templates we have on the shared solidusio/.github repository.

We can delete this file, I added the same template in the shared repository, with an additional
section for releasing an extension. 

When this template is used, it will be an author responsibility to pick the right part of the template
(deleting the rest of the file), depending on if the release is needed for an extension or core.