Add a basic ui/table component used in the products/index component
This is a very basic component that can be used to render a table with
a header and a body. It's not very flexible yet, but it's a start.

The `build_for` method is used to build the table with a basic DSL. It
can be used like this:

    table = SolidusAdmin::UI::Table.build_for(collection) do |t|
      t.column(:name) { |product| product.name }
      t.column(:price) { |product| product.price }
    end

    <%= render table %>

The `column` method accepts a symbol that will be used as the header
along with `human_attribute_name` to get the header text. It also
accepts a block that will be used to render the cell content.

The is a small caveat with how the block is passed to the `build_for`
method. If used in conjunction with `render` the block will be passed to
the `render` method instead of the `build_for` method. To avoid this
problem, the block can be passed with curly braces, or parentheses
should be used to wrap the `render` method call argument.

Co-Authored-By: Rainer Dema <rainer.dema@gmail.com>