Add a basic ui/badge component for the product status column
This is a basic implementation of the ui/badge component. It accepts a
`color` prop and renders a badge with the given color.

The size of the badge is fixed to `M` for now. but it will be possible
to make it dynamic in the future if the design will require other sizes.

Icons are not supported yet, since we don't need them for the product
listing page.

The component is used in the product listing page to display the
`available` / `discontinued` badge in the status column.

Co-Authored-By: Rainer Dema <rainer.dema@gmail.com>