Set up nested main nav items for the admin
We need to add both a `children` and a `top_level` attributes:

- `children` allows us to easily retrieve all the children for a given
  item.
- `top_level` allows us to create different styles for the top level
  items.

`children` can't be used alone, as some of the top level items won't
have children (e.x. users).

We could have used a `parent` attribute instead of both `children` and
`top_level`, but that would have made the implementation more complex,
as we would have to modify the state of the given parent item to
easily retrieve its children. Instead, thanks to the `#with_children`
method we make both attributes private in nature.

We take the ocasion to start testing both the `MainNavItem` and the
item components, as they are now more complex.