Handle active admin menu items from admin main navigation
When it comes to styling, an active menu item has two considerations:

- It's highlighted.
- It's expanded.

We hide nested menu items in any other case.

We consider a menu item to be active if it's generated path matches the
base path (i.e. the path without any query parameters) of the current
request path.