Use remix icons for admin's main navigation items
The images used for the main navigation are part of the Remix icon set
[1]. It's enough to give the freedom to the user to choose an icon from
that set, instead of having to add new icons to the project.

We prefer using a SVG sprite instead of a font because it's easier to
setup in our rails + tailwind environment (we don't need to use a
different CSS file nor deal with npm or sprockets).

[1] - https://remixicon.com/