ResourceController: Allow overriding routes proxy
Within Solidus, we always want to use the `spree` routes proxy as that
is where all routes are defined. However, many people use the Resource
controller as a parent controller in their main app, and because of this
they need to override all the URL generation methods in the Resource
Controller.

This allows one to specify the routes proxy in a custom controller as
follows:

```rb
module MyApp
  class WidgetsController < Spree::Admin::ResourceController
  private

  def routes_proxy
    main_app
  end
end
```

Engine Developers can use this for engines, as well:

```rb
module MyEngine
  class WidgetsController < Spree::Admin::ResourceController
  private

  def routes_proxy
    my_engine
  end
end
```