Use a context object to resolve admin man nav components
Instead of using the magic url helper module with fallbacks to `spree`,
we explicitly pass a context object with access both to the
`solidus_admin` & `spree` proxies. When the `route:` option is given as
a Proc, it's evaluated in that context.

By giving a limited context to the menu item instead of the whole view
context, we avoid exposing too much to an object that doesn't need to
know too many details about the view. That will also prevent us having
to support potential weird usages of the view context, and there's no
need to worry about code injection in case someone uses a main nav item
backed in the database.