Configure admin's typography
We're creating custom Tailwind classes, as our design system contains a
set of well-defined font sizes, line heights, families and weights that
only combine in a few ways. That will help us keep the admin's
typography consistent and easy to maintain.

We're leaving out for now styles for headings, as they're not used yet
and they might require adding a new font family.

Closes #5231