Showcase admin typography classes in lookbook
We don't use a lookbook page because view_component doesn't provide a
shortcut to configure the page paths (as it does with regular previews).
That means we should otherwise hook into the lookbook instance running
in the host application and, if present, add our paths there. With a
single page at this point, it's easier to leverage component previews.