Add account navigation component for the new admin
It's not still fully functional, as authentication is not yet
in place. For now, we use dummy data to show the component.

We're also changing the layout so the sidebar container occupies the full
height of the viewport. That fixes an issue where the sidebar wasn't
filling it due to being in a shorter container.