Add a batch actions form to the ui/table component
The batch action form acts as a hub for row-level checkboxes and
batch-action buttons. The buttons will modify the form's `action` and
`method` attributes so that the list of `id`s will be sent to the
desired action.

The action will receive an array of `id`s inside `param[:id]`, so any
controller action just needs to be adjusted so that it will act on a
collection instead of a single record.

The batch action button needs to also update the forgery protection
token because it depends on the form's action and method, since those
will be overwritten by the buttons, it needs to do the same for the
form authenticity token.

For more information on the `<button>` element's capabilities for
submitting forms see:

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#attributes