Add automatic testing for accessibility in the new admin
We use the axe set of gems [1] to test for automatic testing of accessibility
issues in the new admin. This commit adds the necessary gems and configuration
to the admin Gemfile and spec_helper.rb files. It also adds a test to the
products listing.

As accessibility testing requires javascript, which makes tests slow,
we'll need to be selective about where we add these tests.

[1] - https://github.com/dequelabs/axe-core-gems

Closes #5132