Fix the ui/checkbox component style when focused & checked
The default `background-color: currentColor` was interfering with the
checkbox background color setting it to the same color as the
background.

Setting the background color for all focus combination fixed the
problem.

The focus ring was also tuned to match the provided designs.

The configuration is set to avoid setting form styles in the global
CSS.