Fix Turbo issue with inconsistent form identifiers in Turbo frame
The `batch_actions_form_id` was generating a unique ID with
`SecureRandom.hex` each time it was called.
This led to inconsistencies in form IDs within the turbo frame of our table.

This became problematic when the turbo frame was updated: The form,
identified by a newly generated ID, was not recognized by the table turbo
frame which knew only the original ID. Consequently, upon a form submission
with an action, the payload didn't contain the expected record IDs.

To address this, we've eliminated the random element in ID generation.
We now generate a static ID using `"#{stimulus_id}--batch-actions"`:
This ensures consistency of form ID within the Turbo-frame, enabling Turbo
to handle form submissions and responses correctly.

However, the ideal solution would be to pass the IDs externally, thereby
decoupling the ID generation from the table component and delegating this
responsibility to the page rendering the component.
This approach would also prevent any potential issues that might arise from
having multiple tables on the same page with identical IDs.
This improvement is something we aim to incorporate in the future updates.