Add external id param to enhance table component flexibility
In order to resolve issues with ID inconsistencies and improve the flexibility of our table component, we've introduced a mechanism to pass external IDs to the table component.

This change has allowed us to use a consistent ID across multiple elements
within the table including the batch action form, Turbo-frame, and search form.
By passing the ID externally, we've ensured that the page rendering
the component has full control over ID generation.

This not only resolves the previous issue of Turbo not recognizing the new
identifiers, but it also prevents potential conflicts when multiple tables
are present on the same page.