Add text_area component for the admin
As the logic and styling for the associated label, hint and error
messages are the same as for the text_field component, we extract it
into a couple of new components: one for the label and one for the
guidance (hint and error messages). These new components are marked as
private, as they are not meant to be used directly (at least for now).

The original text_field component and the new text_area compose the new
components but keep the responsibility of how to render the form
inputs and which classes or attributes to use. We consider this
approach more flexible than, for instance, using inheritance. Although
it requires more code than just sharing a bunch of common styles, we
keep them decoupled so they can change independently.

We stop short of creating another form element component wrapping new
extracted pure input and textarea components. We can do that in the
future if we need to.

Ref. #5329