Add `Only show complete orders` filter
Changes include:
- Addition of new filter in `filters` method in the orders index component,
  now includes an additional filter for 'only_show_complete_orders'.

Please note that this is a special case of filters in the context of
transitioning from the old to the new orders index page.
This change ensures that the feature set between the old and new orders index
page remains consistent until we introduce the new filter component.