Add toggletip component
The toggletip component is composed of an info icon which, when clicked
or focused+enter, displays a bubble with a message.

The CSS to correctly display the toggletip and its speech-bubble-style
arrow is quite complex, so we leave the maths inlined to better
understandings of the values.

We use a "button" role on the info icon instead of surrounding it with
an actual button element because we found the latter to break the
positioning of the speech bubble.

For the chosen name and a11y concerns, we've taken inspiration from
https://inclusive-components.design/tooltips-toggletips/#inclusivetoggletips.
TL;DR: it's not good to use a regular "aria-describedby" attribute, as
that would make the text read out in advance. Instead, we build a
clickable control that renders the text on a live region so users are
notified. For progressive enhancement concerns, we render the text
upfront. We use JS to hide it and then add the status role to get
notifications from that moment on.