Fix the legacy switcher not working on different paths
As the cookie was set relative to the path where the switcher was
clicked, it created as many cookies as paths where the switcher
was clicked. That was causing the switcher to not work when a previous
cookie was also encompassing the current path.

To reproduce (in Chrome):

1. Initiate the sandbox application (new admin is on by default)
2. Go to the product creation page. It falls back to the legacy admin
but the switcher is still on the new admin.
3. Click on the switcher to go back to the legacy admin. It works. The
cookie is set for `/admin/products` to `false`.
4. Go to the product listing. Switch again to the new admin. It sets the
cookie for `/admin` to `true`, but as the `/admin/products` cookie
also applies and takes precedence, the new route is not taken into account.