Convert action_name to symbol for CanCanCan authorization
Fix CanCanCan authorization by converting action_name to symbol.
In the context of implementing permissions in the Solidus Demo, we identified
that `action_name` in string format doesn't align with
CanCanCan's `authorize!` method expectations.
It needs to be in symbol format to be processed correctly.
This change ensures our authorization logic functions as intended
also for the new solidus admin views.