Add Facade objects for orders
For calculating and retrieving intermediate amounts on line items,
orders, shipments, and shipping rates, we need some kind of value object
that holds a reference to the order and behaves more or less like the
real things.

I've considered using Ruby Refinements instead of Delegators for this,
but they are clunky in that they need to be referenced in every file
using them.