Skip ActionCable for Dummy Apps
We don't really use ActionCable for anything as far as I know. Having it
makes installing Turbo/Stimulus in extensions harder.

The installer for Turbo will look for a Gemfile in `Rails.root`, which is not
there in case of the dummy app, and fail. If, however, `config/cable.yml` can't be
found, the Turbo installer succeeds with a message that Turbo is
installed, just without the broadcasting feature (which is alright for
most apps I believe).