Display polyvalent rules depending on context
If we specify our product rule to be also applicable to line items, this
rule must be displayed in the list of rules applicable to line items.
However, we don't want to display a form that's already on the page
twice. This commit accomplishes that by displaying, depending on the
context the form is displayed in, either a form or a simple display of
the currently chosen options.