Add row URL proc and standardize parameter naming `ui/table` component
This commit introduces the capability of navigational rows within
the `ui/table` component through the addition of a `row_url` proc.
The `fade_row_proc` parameter is renamed to `row_fade` to maintain a consistent
naming style throughout the component.

The `row_url` proc accepts a row object as a parameter and returns a URL string.
This URL is utilized to redirect the user when a particular row is clicked.
However, this navigation is deliberately disabled in "batch" mode to maintain the
row selection in that mode.

Example of usage:
```ruby
  row_url: ->(order) { spree.edit_admin_order_path(order) }
```