Implement `use_shipping` to sync addresses in Order
Add use_shipping attribute with syncing logic to `Spree::Order` model.
Similar to the existing `use_billing` feature, a `use_shipping` attribute
has been introduced.
This addition enables the automatic assignment of the shipping address
to the billing address when the `use_shipping` flag is set.
The `before_validation` callbacks ensure that the addresses are
synchronized, thereby streamlining the address management process
in the new admin interface.