Enable sortable in `ui/table` component only for default scope
Introduced a fix to conditionally enable the sortable functionality in 
the `ui/table` component.

This enhancement activates sorting options only when the default scope
is applied, which is assumed to align with `Model.all`.

This approach is taken as a precautionary measure to prevent issues in 
data
positioning and sorting when other, more complex scopes are applied.