Allow to choose a custom routes' mount point during install
Since we moved the frontend as an injectable templating system,
there are some manual steps to do to make the application work
when a user wants to change the mount point of where Solidus is
installed.

With this change, along with solidusio/solidus_starter_frontend#359,
we allow the user to select the mount point during the install, and
all the changes required are handled by the install script.