Replace explicit shipping_address and billing_address attributes
Due to deprecations, with a previous commit we replaced `alias_attribute`
calls for these two attributes with `alias_method`, but the latter doesn't
fully cover the previous behavior.

When `shipping_address` or `billing_address` are passed as attributes,
the error `ActiveModel::UnknownAttributeError` is now raised. For this
reason, we're starting to replace them with the right association name.

We may want to revisit the possibility to still use `billing_address` and
`shipping_address` as providing these aliases doesn't seem to be fully
viable anymore.