Provide a prebuilt solidus_admin CSS file
This configuration is a vanilla TW build removing the need for
temporary files and ERB tags. Also relegates the dependency on
tailwindcss-rails only to the development of the gem itself.

By relying on a known target path for the build it will be easy to
setup a local TW build that will override that path in the host Rails
application, leaving the implementors free to go with whatever JS/CSS
builder they want.