Remove unused action in controller callbacks
With Rails 7.1, callbacks only work with actions that are defined
on the controller. This check allowed us to notice we had this some
no more used actions.

This commit also enable the new configuration in the dummy app so we
will be able to spot new errors. The configuration needs to be defined
explicitely because it's not part of the load_defaults but it's copied
in the dummy application directly.

Took advantage of the change for adding another 7.1-only configuration
to the dummy app:

  config.action_dispatch.show_exceptions = :none