Add configuration option: `recalculate_complete_orders`
This gem allows recalculating complete orders until they are shipped,
because Solidus allows changing an order's contents until the order is
shipped.

This change adds a configuration option that disables the
recalculation of orders after they are completed. Some shops make
changes to existing promotions, and expect that this does not change
complete, but unshipped orders.