Theme selection for Solidus Admin: Use spree routing proxy
When using the navigation from a gem that isolates its namespace in
such a way that the default route helper is not spree, this partial
fails to render with an undefined method error.

This can easily be fixed by prefixing the route helper with Solidus' routing proxy.

This is the same problem and solution as #5599, just for the theme
selection partial that is used when using the current admin (rather than
the new one).

(cherry picked from commit f4c96a8ba8cbd7f2dafd7548aa1f0935cf4697e2)