Add extension point for promotion source types
This is an extension point to tell adjustments about what type of
adjustments they are. For the new solidus_friendly_promotions gems, we
need a different type from the `Spree::PromotionAction`. This is not
part of the promotion configuration object, because it is very much
thinkable to have historical adjustments with promotion sources from a
different promotion extension.