Create Spree::SimpleOrderContents
Spree::OrderContents interacts heavily with the current promotion
system. This adds a new `Spree::SimpleOrderContents` class that we can
use once all the core functionality and specs for the legacy promotion
system are migrated to `solidus_legacy_promotions`.

We're keeping `Spree::OrderContents` though, with the intention of
moving it to `solidus_legacy_promotions` in the future.