Add missing product option types
Sample products data was reworked with solidusio/solidus/pull/5437.
After these changes, some products with variants remained without
option types, meaning that their variants remained not visible and
selectable on the FE, and in the admin it was not possible to add
new ones as the relevant button was hidden.