Make Spree::Deprecator available in unit tests
It seems we want to be able to test files without loading all of Solidus.
Unfortunately, we do load Solidus' configuration, and this PR needs the
configuration object to be able to deprecate extension points. What this
commit does is it allows loading a module that defined
`Spree.deprecator` and loads that module in both `spree/core.rb` and the
`spec_helper.rb`.