Test promotion totals changing
We would rather keep `Spree::Order#ensure_promotions_eligible` in core,
as what it does is re-run the promotion adjuster class and make sure the
amounts are still correct. This should be fine for every promotion
system. However, since the Null adjuster doesn't actually do anything,
we need to stub what it does in the spec.