Wrap order email modal with turbo frame tag
This way, the modal can be loaded async into the current page.

Please note that the order page rendering, which is discarded on
turbo-frame calls, needs to be moved to the bottom, as it includes
an empty turbo frame tag with same ID that would interfere with the
modal rendering.