Enable turbo frame links in URL params
`Turbo.visit` already supports turbo frames, so we just need a way
to pass the information from the table HTML to the `visit` call.

This implementation, although arbitrary and possibly questionable,
has the merit of being simple, meaning that enabling turbo frames
in a table row requires just to update the URL to include the
`_turbo_frame` param. This param is removed from the URL, so it
won't have any possible further side effect in the application.